feat: add Discord controller
