+++ html/admin-lord-sedwyn.php
@@ -6,7 +6,7 @@ $session = require(\Kickback\SCRIPT_ROOT . "/api/v1/engine/session/verifySession
 require("php-components/base-page-pull-active-account-info.php");
 
 use Kickback\Services\Session;
-use Kickback\Backend\Controllers\SocialMediaController;
+use Kickback\Backend\Controllers\DiscordController;
 use Kickback\Backend\Config\ServiceCredentials;
 if (!Session::isMagisterOfTheAdventurersGuild())
 {
@@ -83,7 +83,7 @@ $webhookURL = $kk_credentials["discord_api_url"] . '/' . $kk_credentials["discor
                 if ($_SERVER["REQUEST_METHOD"] === "POST" && isset($_POST["msg"])) {
                     $msg = trim($_POST["msg"]);
                     if ($msg !== "") {
-                        SocialMediaController::sendDiscordWebhook($msg);
+                        DiscordController::sendWebhook($msg);
                         echo '<div class="alert alert-success mt-3" role="alert">';
                         echo '<strong>Message sent:</strong> ' . htmlspecialchars($msg);
                         echo '</div>';
