+++ html/api/v1/engine/discord/unlink.php
@@ -1,11 +1,11 @@
 <?php
 require(__DIR__.'/../../engine/engine.php');
 
-use Kickback\Backend\Controllers\SocialMediaController;
+use Kickback\Backend\Controllers\DiscordController;
 use Kickback\Services\Session;
 
 OnlyPOST();
 
 $account = Session::requireDiscordLinked();
-return SocialMediaController::unlinkDiscordAccount($account);
+return DiscordController::unlinkAccount($account);
 ?>
